+++ NEWS.md
@@ -1,3 +1,5 @@
+# pmxhelpr (development version)
+
 # pmxhelpr 0.3.5
 
 * Add option to pass the same dataset variable to `TIME` and `NTIME` attributes in `time_vars` for all functions